requirements.txt
torch
accelerate
transformers
gradio
langdetect
pyaudio
filetype
pyannote.audio
python-dotenv